# src/main/java/com/ibagroup/carrental/dto/user/UserDto.java
package com.ibagroup.carrental.dto.user;

import lombok.Data;

@Data
public class UserDto {

    private Long id;
    private String userName;
    private String firstName;
    private String lastName;
    private String email;
    private String password;
    private String phone;

}